srcCodeAPIs/alchemy/requirements.txt
flask
flask_sqlalchemy
pyjwt
flask-swagger-ui
